requirements.txt
requests==2.26.0
beautifulsoup4==4.9.3
python-dotenv==0.19.0
streamlit>=1.24.0
langchain>=0.0.267
langchain-google-genai==0.0.11
google-generativeai>=0.3.0
duckduckgo-search>=3.9.0
arxiv 